+++ js/data.js
@@ -1,0 +1,26 @@
+//ДАННЫЕ
+
+const NAMES = [
+  'Иван',
+  'Вера',
+  'Петя',
+  'Лера',
+  'Женя',
+  'Сева',
+  'Юра',
+  'Виталя',
+  'Аркадий',
+  'Влад',
+  'Анна',
+  'Дмитрий',
+  'Анастасия',
+];
+
+const MESSAGES = 'Всё отлично!\n' +
+  'В целом всё неплохо. Но не всё.\n' +
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.\n' +
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.\n' +
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.\n' +
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!\n';
+
+export {NAMES, MESSAGES};
